Refactored Modix.Data.Models namespace.
