+++ Modix.Data/Models/Core/ChannelLimitEntity.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Modix.Data.Models
+namespace Modix.Data.Models.Core
 {
     public class ChannelLimitEntity
     {
