+++ Modix.Data/Models/Core/DiscordGuildEntity.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Modix.Data.Models
+namespace Modix.Data.Models.Core
 {
     public class DiscordGuildEntity
     {
