+++ Modix.Data/Models/Core/GuildConfig.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Modix.Data.Models
+namespace Modix.Data.Models.Core
 {
     public class GuildConfig
     {
