+++ Modix.Data/Models/Moderation/InfractionEntity.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
+using Modix.Data.Models.Core;
+
 namespace Modix.Data.Models.Moderation
 {
     /// <summary>
