+++ Modix.Data/Models/Promotion/PromotionCommentEntity.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace Modix.Data.Models
+namespace Modix.Data.Models.Promotion
 {
     public enum PromotionSentiment
     {
