+++ Modix.Data/ModixContext.cs
@@ -4,8 +4,9 @@ using System.Threading.Tasks;
 
 using Microsoft.EntityFrameworkCore;
 
-using Modix.Data.Models;
+using Modix.Data.Models.Core;
 using Modix.Data.Models.Moderation;
+using Modix.Data.Models.Promotion;
 
 namespace Modix.Data
 {
