+++ Modix.Data/Repositories/UserRepository.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-using Modix.Data.Models;
+using Modix.Data.Models.Core;
 
 namespace Modix.Data.Repositories
 {
