+++ Modix.Services/Promotions/IPromotionRepository.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Discord.WebSocket;
-using Modix.Data.Models;
+using Modix.Data.Models.Promotion;
 
 namespace Modix.Services.Promotions
 {
