+++ Modix/ModixBot.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
 using Discord.WebSocket;
-using Modix.Data.Models;
+using Modix.Data.Models.Core;
 using Serilog;
 using Microsoft.Extensions.DependencyInjection;
 using Modix.Services.Quote;
