+++ Modix/Modules/PromotionModule.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
 using Discord.WebSocket;
-using Modix.Data.Models;
+using Modix.Data.Models.Promotion;
 using Modix.Services.Promotions;
 
 namespace Modix.Modules
